server/requirements.txt
asgiref==3.5.0; python_version >= "3.7"
dj-rest-auth==2.1.7
django-cors-headers==3.7.0
django==3.2.16
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.17.1
pytz==2022.1
setuptools==62.1.0; python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.5"
